rearrange directories in a possibly misguided effort to follow some sort of convention
